\name{CNS}
\alias{CNS}
\docType{data}
\title{
Central Nervous System (CNS) cancer data set.
}

\description{
A list object includes the measurement of the gene expression in 7128 genes and 60 samples and the clinical risk factors which are Chang stage (nominal), sex (binary), age (nominal), chemo Cx (binary), chemo VP (binary). 21 patients died (labeled 0) and 39 patients survived (labeled 1) within 24 months after the treatment.
}


\references{
Pomeroy, SL. and Tamayo, P. and Gaasenbeek, M. and Sturla, LM. and Angelo, M. and McLaughlin, ME. and Kim, JY. and Goumnerova, LC. and Black, PM. and Lau, C. and Allen, JC. and Zagzag, D. and Olson, JM. and Curran, T. and Wetmore, C. and Biegel, JA. and Poggio, T. and Mukherjee, S. and Rifkin, R. and Califano, A. and Stolovitzky, G. and Louis, DN. and Mesirov, JP. and Lander, ES. and Golub, TR. Prediction of Central Nervous System Embryonal Tumour Outcome Based on Gene Expression. \emph{Nature}., 415(6870):436-442.
}

\examples{
data(CNS)
names(CNS)
}

